+++ apiListener/package.json
@@ -1,0 +1,19 @@
+{
+  "name": "apilistener",
+  "version": "1.0.0",
+  "description": "",
+  "main": "main.js",
+  "scripts": {
+    "test": "echo \"Error: no test specified\" && exit 1",
+    "serve": "node main.js"
+  },
+  "repository": {
+    "type": "git",
+    "url": "http://192.168.1.160:3000/michael/Pi-Hole-Dashboard"
+  },
+  "author": "",
+  "license": "ISC",
+  "dependencies": {
+    "express": "^4.21.2"
+  }
+}
