+++ backend/main.js
@@ -1,4 +1,4 @@
-const express = require("express");
+// const express = require("express");
 const WebSocket = require('ws');
 const http = require('http');
 const { parseMessage } = require('./utils.js')
@@ -9,8 +9,8 @@ const { writeMultiLogs } = require('./influxConnection.js');
 const server = http.createServer();
 const wss = new WebSocket.Server({ server });
 const clients = new Set();
-const app = express();
-app.use(express.json());
+// const app = express();
+// app.use(express.json());
 
 // Connect new clients
 wss.on('connection', (clientSocket) => {
@@ -41,36 +41,36 @@ wss.on('connection', (clientSocket) => {
    });
 });
 
-const PORT = process.env.PI_HOLE_API_PORT || API_LISTEN_PORT;
+// const PORT = process.env.PI_HOLE_API_PORT || API_LISTEN_PORT;
 
-const disablePiHole = async (url, apiSendDisableMinutesTokenPart1, minutes, apiAndAuth, key) => {
-   const seconds = minutes * 60;
-   const response = await fetch(url + apiSendDisableMinutesTokenPart1 + seconds + apiAndAuth + key);
-   if (!response.ok) {
-      throw new Error('Network response was not ok ' + response.status);
-   }
-};
+// const disablePiHole = async (url, apiSendDisableMinutesTokenPart1, minutes, apiAndAuth, key) => {
+//    const seconds = minutes * 60;
+//    const response = await fetch(url + apiSendDisableMinutesTokenPart1 + seconds + apiAndAuth + key);
+//    if (!response.ok) {
+//       throw new Error('Network response was not ok ' + response.status);
+//    }
+// };
 
-const disablePis = async (minutes = 10) => {
-   try {
-      disablePiHole(PI1_URL, apiSendDisableMinutesTokenPart1, minutes, apiAndAuth, PI1_KEY);
-      disablePiHole(PI2_URL, apiSendDisableMinutesTokenPart1, minutes, apiAndAuth, PI2_KEY);
-   } catch (error) {
-      console.error('Error sending data:', error);
-   }
-};
+// const disablePis = async (minutes = 10) => {
+//    try {
+//       disablePiHole(PI1_URL, apiSendDisableMinutesTokenPart1, minutes, apiAndAuth, PI1_KEY);
+//       disablePiHole(PI2_URL, apiSendDisableMinutesTokenPart1, minutes, apiAndAuth, PI2_KEY);
+//    } catch (error) {
+//       console.error('Error sending data:', error);
+//    }
+// };
 
-app.get(`/${API_DISABLE_PI}`, async (req, res) => {
-   const disabled = req.query.disable || true;
-   if (disabled) {
-      await disablePis();
-   }
-   res.sendStatus(200);
-});
+// app.get(`/${API_DISABLE_PI}`, async (req, res) => {
+//    const disabled = req.query.disable || true;
+//    if (disabled) {
+//       await disablePis();
+//    }
+//    res.sendStatus(200);
+// });
 
-app.listen(PORT, () => {
-   console.log(`REST API listening on port ${PORT}`);
-});
+// app.listen(PORT, () => {
+//    console.log(`REST API listening on port ${PORT}`);
+// });
 server.listen(WEBSOCKET_SERVER_PORT, '0.0.0.0', () => {
    console.log(`WebSocket server running on ws://backend:${WEBSOCKET_SERVER_PORT}`);
 });
